qemu bit working again.  At least it creates a log...

+++ hw/ide.c
@@ -769,7 +769,7 @@ static void ide_sector_read(IDEState *s)
 
 	// from HD to IO_BUFFER
 	// (op,from_addr,to_addr,size)
-	iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, HD_BASE_ADDR + sector_num*512, s->io_buffer, n*512);
+	iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, HD_BASE_ADDR + sector_num*512, (uint64_t) s->io_buffer, n*512);
         ret = bdrv_read(s->bs, sector_num, s->io_buffer, n);
         ide_transfer_start(s, s->io_buffer, 512 * n, ide_sector_read);
         ide_set_irq(s);
@@ -818,6 +818,7 @@ static int dma_buf_rw(BMDMAState *bm, int is_write)
 	      // (op,from_addr,to_addr,l)
 	      //	      iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, IO_BUFFER_BASE_ADDR + s->io_buffer_index, bm->cur_prd_addr,  l);
 
+	      // no need for iferret logging!
 	      // this one should just get handled by cpu_physical_memory_rw...
                 cpu_physical_memory_write(bm->cur_prd_addr,
                                           s->io_buffer + s->io_buffer_index, l);
@@ -833,6 +834,7 @@ static int dma_buf_rw(BMDMAState *bm, int is_write)
 	      //	      iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, bm->cur_prd_addr, IO_BUFFER_BASE_ADDR + s->io_buffer_index, l);	
 
 
+	      // no need for iferret logging.
 	      // this one should just get handled by cpu_physical_memory_rw...
                 cpu_physical_memory_read(bm->cur_prd_addr,
                                           s->io_buffer + s->io_buffer_index, l);
@@ -892,7 +894,7 @@ static void ide_read_dma_cb(void *opaque, int ret)
     //    IFLW_HD_TRANSFER(HD_BASE_ADDR + sector_num*512, IO_BUFFER_BASE_ADDR, n*512);
 
     // HD -> IO_BUFFER
-    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, HD_BASE_ADDR + sector_num*512, s->io_buffer, n*512);
+    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, HD_BASE_ADDR + sector_num*512, (uint64_t) s->io_buffer, n*512);
    bm->aiocb = bdrv_aio_read(s->bs, sector_num, s->io_buffer, n,
                               ide_read_dma_cb, bm);
 }
@@ -927,7 +929,7 @@ static void ide_sector_write(IDEState *s)
     //    IFLW_HD_TRANSFER(IO_BUFFER_BASE_ADDR, HD_BASE_ADDR + sector_num*512, n*512);
 
     // IO_BUFFER -> HD
-    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, s->io_buffer, HD_BASE_ADDR + sector_num*512, n*512);
+    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, (uint64_t) s->io_buffer, HD_BASE_ADDR + sector_num*512, n*512);
   //search_buf_for_pattern((char*)s->io_buffer,n*512);
     ret = bdrv_write(s->bs, sector_num, s->io_buffer, n);
     s->nsector -= n;
@@ -1004,7 +1006,7 @@ static void ide_write_dma_cb(void *opaque, int ret)
 
 
     // IO_BUFFER -> HD
-    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, s->io_buffer, HD_BASE_ADDR + sector_num*512, n*512);
+    iferret_log_info_flow_op_write_884(IFLO_HD_TRANSFER, (uint64_t) s->io_buffer, HD_BASE_ADDR + sector_num*512, n*512);
     bm->aiocb = bdrv_aio_write(s->bs, sector_num, s->io_buffer, n,
                                ide_write_dma_cb, bm);
 }
@@ -2341,7 +2343,7 @@ static void ide_data_writew(void *opaque, uint32_t addr, uint32_t val)
 
     // (op,to,size)
     // ? -> IO_BUFFER 
-    iferret_log_info_flow_op_write_84(IFLO_HD_TRANSFER_PART2, s->io_buffer + (s->data_ptr - s->data_ptr_base), 2);
+    iferret_log_info_flow_op_write_84(IFLO_HD_TRANSFER_PART2, (uint64_t) s->io_buffer + (s->data_ptr - s->data_ptr_base), 2);
     p = s->data_ptr;
     *(uint16_t *)p = le16_to_cpu(val);
     p += 2;
@@ -2363,7 +2365,7 @@ static uint32_t ide_data_readw(void *opaque, uint32_t addr)
 
     // (op,from)
     // IO_BUFFER -> 
-    iferret_log_info_flow_op_write_8(IFLO_HD_TRANSFER_PART1, s->io_buffer + (s->data_ptr - s->data_ptr_base));
+    iferret_log_info_flow_op_write_8(IFLO_HD_TRANSFER_PART1, (uint64_t) s->io_buffer + (s->data_ptr - s->data_ptr_base));
     /*
     printf("IO_BUFF starts at %llu and offset is %u\r\n",IO_BUFFER_BASE_ADDR,s->data_ptr - s->data_ptr_base);
     printf("phys_ram_size is %llu\r\n",(unsigned long long)phys_ram_size);
@@ -2385,7 +2387,7 @@ static void ide_data_writel(void *opaque, uint32_t addr, uint32_t val)
     //    IFLW_HD_TRANSFER_PART2(IO_BUFFER_BASE_ADDR + (s->data_ptr - s->data_ptr_base), 4);
 
     // -> IO_BUFFER
-    iferret_log_info_flow_op_write_84(IFLO_HD_TRANSFER_PART2, s->io_buffer + (s->data_ptr - s->data_ptr_base), 4);
+    iferret_log_info_flow_op_write_84(IFLO_HD_TRANSFER_PART2, (uint64_t) s->io_buffer + (s->data_ptr - s->data_ptr_base), 4);
     p = s->data_ptr;
     *(uint32_t *)p = le32_to_cpu(val);
     p += 4;
@@ -2403,7 +2405,7 @@ static uint32_t ide_data_readl(void *opaque, uint32_t addr)
     //    IFLW_HD_TRANSFER_PART1(IO_BUFFER_BASE_ADDR + (s->data_ptr - s->data_ptr_base));
 
     // IO_BUFFER -> 
-    iferret_log_info_flow_op_write_8(IFLO_HD_TRANSFER_PART1, s->io_buffer + (s->data_ptr - s->data_ptr_base));
+    iferret_log_info_flow_op_write_8(IFLO_HD_TRANSFER_PART1, (uint64_t) s->io_buffer + (s->data_ptr - s->data_ptr_base));
     /*
     printf("IO_BUFF starts at %llu and offset is %u\r\n",IO_BUFFER_BASE_ADDR,s->data_ptr - s->data_ptr_base);
     printf("phys_ram_size is %llu\r\n",(unsigned long long)phys_ram_size);
